Default to returning u64 in Rust solution template
There's no performance penalty that I'm aware of, and some solutions
need it.

+++ rust/template.rs
@@ -1,11 +1,11 @@
 const EXAMPLE: &str = include_str!("example.txt");
 const INPUT: &str = include_str!("input.txt");
 
-fn part1(input: &str) -> u32 {
+fn part1(input: &str) -> u64 {
     0
 }
 
-fn part2(input: &str) -> u32 {
+fn part2(input: &str) -> u64 {
     0
 }
